Update README.md
In this project I implmented a dynamic array where the elements sand, water and metal will be stored. They each have different weights and respond differently when adding more of that element or changing its enviroment. It is entertaining to see how the grid can make the sand and the water move while the metal collapses if you make the grid smaller.